Add SummaryStats component
